Update pagination for Category Interface

+++ src/main/java/com/dev/shoeshop/controller/admin/AdminCategoryController.java
@@ -26,7 +26,7 @@ public class AdminCategoryController {
     }
 
     @GetMapping("/update/{id}")
-    public String updateCategoryPage(@PathVariable Integer id) {
+    public String updateCategoryPage(@PathVariable Long id) {
         Category category = categoryService.getCategoryById(id);
         return "admin/categories/category-edit";
     }
